requirements.txt
duckdb~=1.1.4.dev4108
streamlit~=1.41.1
pandas~=2.2.1
DateTime~=5.5
fastapi~=0.115.6
boolean~=1.1.0
python-dateutil~=2.9.0.post0
fastparquet~=2024.11.0
pyarrow~=18.1.0
stringcase~=1.2.0
prompt_toolkit~=3.0.48
streamlit-aggrid~=1.0.5
requests~=2.32.3
altair~=5.5.0
EigenLedger~=2.1.6
matplotlib~=3.10.0
seaborn~=0.13.2
plotly~=6.0.0
numpy~=1.26.4
statsmodels~=0.14.4
scikit-learn~=1.6.1
pandas-datareader~=0.10.0
yfinance~=0.2.54
arch~=7.2.0
scikit-optimize~=0.10.2